+++ frontend/src/components/modal/modal.css
@@ -1,0 +1,29 @@
+/* MODAL STYLING */
+.modal-background {
+  position: fixed;
+  top: 0;
+  bottom: 0;
+  right: 0;
+  left: 0;
+  background: rgba(0, 0, 0, 0.7);
+  z-index: 10;
+}
+
+.modal-child {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
+
+
+.modal-close{
+  font-size: 18px;
+  font-weight: normal;
+  margin-bottom: 0px;
+  border: none;
+  color: grey;
+  text-align: left;
+  position: fixed;
+  top: 45px;
+}
