+++ frontend/src/components/nav/nav.css
@@ -1,0 +1,103 @@
+.home-logo-comp{
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+  margin-left: 15px;
+  /* padding: 10px 50px; */
+}
+
+.home-logo{
+  width: 50px;
+  height: 50px;
+  margin: 10px 30px;
+}
+
+
+.nav-search{
+  width: 300px;
+  height: 35px;
+  background-color: none;
+  border: 2px solid rgba(163, 3, 163, 0.486);
+  margin-left: 30px;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  padding: 0px 5px;
+}
+
+.search-input{
+  width: 250px;
+  height: 100%;
+  padding: 0px 10px;
+  margin-left: 10px;
+  font-size: 18px;
+  outline: none;
+  border: none;
+  background: none;
+  color: white;
+}
+
+.search-input:focus{
+  outline: none;
+  /* background-color: rgba(255, 255, 255, 0.561) */
+}
+
+.search-icon {
+  color: white;
+  font-size: 26px;
+}
+
+.nav-bar-list{
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-around;
+  margin-right: 15px;
+  /* border: 1px solid red; */
+  height: 80px;
+  width: 7%;
+}
+
+.nav-link{
+  text-decoration: none;
+  color: black;
+  width: 95%;
+  height: 50%;
+  padding: 0% 5%;
+  display: flex;
+  align-items: center;
+}
+
+.nav-link:hover{
+  color: white;
+  background-color: purple;
+}
+
+.user-icon-div{
+  height: 100%;
+  width: 100px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.user-icon{
+  font-size: 40px;
+}
+
+.user-icon:hover{
+  color:purple
+}
+
+.user-options {
+  position: absolute;
+  top: 80px;
+  right: 0;
+  width: 120px;
+  height: 80px;
+  background-color: white;
+  border: 1px solid black;
+  border-bottom-left-radius: 5px;
+  border-bottom-right-radius: 5px;
+}
