+++ frontend/src/components/nav/nav_bar.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Link } from "react-router-dom";
+import './nav.css';
 
 class NavBar extends React.Component{
     constructor(props){
